sqlalchemy_tools/requirements.txt
certifi==2021.10.8
greenlet==1.1.2
psycopg2-binary==2.9.3
python-dotenv==0.19.2
SQLAlchemy==1.4.31
